refactor: Memo 내 lastModified Double 타입으로 변경

+++ CloudNotes/CloudNotes/Model/Memo.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Memo: Decodable {
     let title: String
     let body: String
-    let lastModified: Int
+    let lastModified: Double
     
     enum CodingKeys: String, CodingKey {
         case title
